More de-linting
Since yamllint now supports disabling / re-enabling checks on the fly we
can now consider adding it as part of our validation for JJB. However,
before we do that, we want to get everything currently in JJB passing
it. This disables line-length checking on the lines that need it to be
disabled on.

Change-Id: I8998a710eaedc16a5469f4253362003c0db6da4e
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>