Fix the maven settings being used
Our infrastructure requires that each project have custom maven settings
for the project itself as well as a global maven configuration file for
the repository configurations. This updates the templates to include the
needed configuration as well as setting the configuration definition for
each of the projects that already have jobs defined.

Change-Id: Ic0767b5253e6cbaffc0c90cbd60e2063a2f6b3fb
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>