Initial thirdparty-winery build specification
Get a job running that does the build of winery. This does not deploy to
Nexus yet as that that still needs a little bit of work / testing given
that they source repo has all objects configured as SNAPSHOT and a way
to deal with building release versions needs to be handled.

Related JIRA
TOSCA-49
OCS-47

Change-Id: Ia57cab4fde68d4180e10c74e535b6380cb656875
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>