Fix starting point for autorelease staging push
The autorelease maven deployment was brought in from OpenDaylight which
has a separate project for autorelease and thus does all processing at
the $WORKSPACE level. As the autorelease component for integration is
part of the integration repository as a subdirectory we need to be aware
of this and reroot any work that needs to happen appropriately.

Issue-id: INT-9

Change-Id: I9393e99436aa1951c2e3128efe782c79663d901a
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>