Start shipping all build logs to log hosting
A new log storage service has been built out to allow logs to be hosted
for up to 6 months instead of the 40 builds or 30 days (which ever is
smaller) that is configured in Jenkins. Jobs will be automatically
updated with with information for where the logs are being stored. An
additional upcoming modification to Jenkins will start pushing this
information into the triggering Gerrit changes as well.

Issue-ID: CIMAN-1

Change-Id: If736eb8acb86791ac75f4916434a05457e897cde
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>